+++ crates/vyxen/src/prelude.rs
@@ -5,3 +5,4 @@ pub use vyxen_input::*;
 pub use vyxen_math::*;
 pub use vyxen_physics2d::*;
 pub use vyxen_renderer::*;
+pub use vyxen_resource::*;