+++ CHANGELOG.md
@@ -1,1 +1,15 @@
-Nothing here yet.
+# Changelog
+
+## Unreleased
+
+### New Features
+
+- Resource loading
+
+### Fixes
+
+- Textures being rendered 180 degrees
+
+## v0.1.0 (24/07/2026)
+
+Initial release